Use graphQL datasource
In this commit, we change the main data source from Firestore to GraphQL. The session data is already merged, so no need to manually merge the talks, rooms and slots.
We deactivated the JSONDataProvider because it was too hard to have the same behavior with such different data sources. We will try to add it later.
The OpenFeedback data is still using Firestore.